requirements.txt
asgiref==3.5.2
astroid==2.12.11
async-timeout==4.0.2
black==22.10.0
click==8.1.3
Deprecated==1.2.13
dill==0.3.5.1
dj-database-url==1.0.0
Django==4.1.2
django-bootstrap-form==3.4
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-on-heroku==1.1.2
django-redis==5.2.0
djangorestframework==3.14.0
flake8==5.0.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pyflakes==2.5.0
pylint==2.15.4
pyparsing==3.0.9
pytz==2022.5
redis==4.3.4
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.5
whitenoise==6.2.0
wrapt==1.14.1
